Begin experiment: sync context to enable breaking execution flow
Right now if you do  in the midst of continuation code
running as the result of a load, control will return to the loader and context,
and any load calls made after the  will not be processed in this run.

These calls may potentially be lost, or another call to
may be required.

This means to find a solution we may need to  any Tasks that we did not make
and continue past any Tasks that did (using  in ) method.
I'm not sure if we can do this with a SynchronizationContext alone, or even with a custom TaskScheduler.
Investigation is required on the TaskScheduler front, but we may need to resort to using a custom
Awaitable type that notifies the loader when it has been awaited, at which point the loader can proceed.