Simplified implementation; added a root loader
A root loader has a parameterless fetch and is useful for kicking off the chain of loaders.
For example, it would be used to resolve the root `episodes` field.

Cases that don't require the usual collect + fire pattern should still be encapsulated
as some kind of loader so that they're integrated into the loader framework, which is necessary
so we can track them and execute subsequent loaders reliably and efficiently to fulfil a request.